Create Assignment No. 2 (A) Bubble sort
write a game development program that implements the Bubble Sort algorithm. The
program will simulate a simple game where the player can input a set of numbers,
and the numbers will be sorted using Bubble Sort to simulate a "level-up" scenario
where the player's scores are sorted in ascending order.